🧬 Migrate the object *_ReferenceObjectParser* into module *_schema_parser* to be clear where it should belong to.
