+++ pymock_api/model/openapi/_base.py
@@ -1,10 +1,9 @@
 from abc import ABCMeta, abstractmethod
-from typing import Any, Dict, List, Union
+from typing import Any, Dict, Union
 
 from ..api_config import _Config
 from ..enums import OpenAPIVersion
 from ._parser_factory import BaseOpenAPISchemaParserFactory, get_schema_parser_factory
-from ._schema_parser import BaseOpenAPISchemaParser
 
 Self = Any
 
@@ -68,50 +67,3 @@ class Transferable(BaseOpenAPIDataModel):
     @abstractmethod
     def to_api_config(self, **kwargs) -> _Config:
         pass
-
-
-ComponentDefinition: Dict[str, dict] = {}
-
-
-def get_component_definition() -> Dict:
-    global ComponentDefinition
-    return ComponentDefinition
-
-
-def set_component_definition(openapi_parser: BaseOpenAPISchemaParser) -> None:
-    global ComponentDefinition
-    ComponentDefinition = openapi_parser.get_objects()
-
-
-class _ReferenceObjectParser:
-    @classmethod
-    def has_schema(cls, data: Dict) -> bool:
-        return data.get("schema", None) is not None
-
-    @classmethod
-    def has_ref(cls, data: Dict) -> str:
-        if cls.has_schema(data):
-            has_schema_ref = data["schema"].get("$ref", None) is not None
-            return "schema" if has_schema_ref else ""
-        else:
-            _has_ref = data.get("$ref", None) is not None
-            return "ref" if _has_ref else ""
-
-    @classmethod
-    def get_schema_ref(cls, data: dict) -> dict:
-        def _get_schema(component_def_data: dict, paths: List[str], i: int) -> dict:
-            if i == len(paths) - 1:
-                return component_def_data[paths[i]]
-            else:
-                return _get_schema(component_def_data[paths[i]], paths, i + 1)
-
-        print(f"[DEBUG in get_schema_ref] data: {data}")
-        _has_ref = _ReferenceObjectParser.has_ref(data)
-        if not _has_ref:
-            raise ValueError("This parameter has no ref in schema.")
-        schema_path = (
-            (data["schema"]["$ref"] if _has_ref == "schema" else data["$ref"]).replace("#/", "").split("/")[1:]
-        )
-        print(f"[DEBUG in get_schema_ref] schema_path: {schema_path}")
-        # Operate the component definition object
-        return _get_schema(get_component_definition(), schema_path, 0)
