+++ test/unit_test/model/openapi/inner_parser.py
@@ -5,9 +5,7 @@ import pytest
 
 from pymock_api.model.enums import OpenAPIVersion, ResponseStrategy
 from pymock_api.model.openapi._base import (
-    _ReferenceObjectParser,
     get_schema_parser_factory_with_openapi_version,
-    set_component_definition,
     set_openapi_version,
     set_parser_factory,
 )
@@ -15,6 +13,8 @@ from pymock_api.model.openapi._parser import APIParameterParser, APIParser
 from pymock_api.model.openapi._schema_parser import (
     OpenAPIV2PathSchemaParser,
     OpenAPIV2SchemaParser,
+    _ReferenceObjectParser,
+    set_component_definition,
 )
 from pymock_api.model.openapi.config import APIParameter
